refactor: moved styles to seperate folder
