--- app/components/IndividualSummoner/IndividualSummoner.jsx
@@ -1,1 +1,0 @@
-IndividualSummoner.jsx
