+++ app/styles/Carousel.css
