+++ app/styles/LiveGame.css
