+++ app/styles/index.jsx
@@ -1,0 +1,1 @@
+index.jsx
